+++ MY_RoleStatistics/src/MY_RoleStatistics_EquipStat.lua
@@ -903,8 +903,10 @@ end
 
 X.RegisterUserSettingsUpdate('@@INIT@@', 'MY_RoleStatistics_EquipStat', function()
 	D.bReady = true
-	D.UpdateSaveDB()
-	D.FlushDB()
+	if not ENVIRONMENT.RUNTIME_OPTIMIZE then
+		D.UpdateSaveDB()
+		D.FlushDB()
+	end
 	D.UpdateFloatEntry()
 end)
 X.RegisterReload('MY_RoleStatistics_EquipStat', function() D.ApplyFloatEntry(false) end)
