+++ MY_RoleStatistics/src/MY_RoleStatistics_TaskStat.lua
@@ -1383,8 +1383,10 @@ function D.UpdateFloatEntry()
 end
 X.RegisterUserSettingsUpdate('@@INIT@@', 'MY_RoleStatistics_TaskStat', function()
 	D.bReady = true
-	D.UpdateSaveDB()
-	D.FlushDB()
+	if not ENVIRONMENT.RUNTIME_OPTIMIZE then
+		D.UpdateSaveDB()
+		D.FlushDB()
+	end
 	D.UpdateFloatEntry()
 end)
 X.RegisterReload('MY_RoleStatistics_TaskEntry', function() D.ApplyFloatEntry(false) end)
